agregar pruebas unitarias
Se agregan las pruebas unitarias para la facturacion de precios precios

+++ src/main/java/com/ceiba/induccion/utils/factura/Factura.java
@@ -75,7 +75,7 @@ public class Factura {
 		}
 	}
 	
-	public float getPrecio() {
+	public int getPrecio() {
 		return precio;
 	}
